server/pyproject.toml
[project]
name = "soma-sophia-server"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "authlib>=1.6.9",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pyright]
typeCheckingMode = "standard"
reportUnknownParameterType = true
venvPath = "."
venv = ".venv"
